Revert "fix: caption and credits line height on ios (edge case)"
This reverts commit cfbbeb184a9797d7f73c343e21588066a0e15dc2.
